requirements.txt
Django==5.0.7
Django==5.1
mysql_connector_repackaged==0.3.1
